Rewriting example
